+++ popkorn-example/src/main/kotlin/cc/popkorn/example/model/R.kt
@@ -1,6 +1,7 @@
 package cc.popkorn.example.model
 
 import cc.popkorn.annotations.*
+import cc.popkorn.core.Injector
 import cc.popkorn.core.model.Empty
 import cc.popkorn.core.model.Environment
 
@@ -23,7 +24,7 @@ class R5(str: String, int: Int, test: Float?, @WithEnvironment("env1") d1: DiC, 
 class R6(@WithEnvironment("env4") d1: DiC, @WithEnvironment("env4") d2: DiC, @Alias("is7") d3: DiC, @Alias("is8") d4: DiC) : R6i
 
 @Injectable
-class R7(tmp: R0i?, d1: Wrapper.DiD, @WithEnvironment("envX") d2: Wrapper.DiD, d3: D10) : R7i
+class R7(injector: Injector, tmp: R0i?, d1: Wrapper.DiD, @WithEnvironment("envX") d2: Wrapper.DiD, d3: D10) : R7i
 
 @Injectable
 @ForEnvironments("env1", "env2", "env3", "env4")