+++ popkorn-example/src/main/kotlin/cc/popkorn/example/model/D.kt
@@ -36,6 +36,15 @@ class D9 : Wrapper.DiD, DiG, DiB
 
 class D10 : Wrapper.DiD
 
+open class D11 : DaI(), DiE
+
+@Injectable(propagation = Propagation.ALL)
+class D12 : D11(), DiF
+
+class D13 : DaJ(), DiF
+
+@Injectable(propagation = Propagation.DIRECT)
+class D14 : DaJ()
 
 @InjectableProvider(scope = Scope.BY_NEW, propagation = Propagation.DIRECT, alias = "is6")
 @ForEnvironments("env1", "env2", "env3")
@@ -57,3 +66,14 @@ class DCustom7 {
 class DCustom8 {
     fun create() = D8()
 }
+
+@InjectableProvider
+@ForEnvironments("env9", "env2")
+class DCustom9 {
+    fun create(d0: DiA): DaJ = D13()
+
+    @ForEnvironments("env2")
+    fun createEnv2(): DaJ = D13()
+
+}
+