+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 Change Log
 ==========
 
-Version 1.4.0 *(XXXXXX)*
+Version 1.4.0 *(2020-XX-XX)*
 -----------------------------
+* New: Add support for abstract classes as if they where interfaces
 * Fix: Add support to all prohibited packages(java.*, javax.*), not only java.lang.*
 * New: Can get the injector at any injectable constructor/method